+++ server/runtime_go_nakama.go
@@ -1650,14 +1650,14 @@ func (n *RuntimeGoNakamaModule) NotificationsSend(ctx context.Context, notificat
 }
 
 // @group notifications
-// @summary Send one in-app notification to all users.
+// @summary Send an in-app notification to all users.
 // @param ctx(type=context.Context) The context object represents information about the server and requester.
 // @param subject(type=string) Notification subject.
-// @param content(type=map[string]interface{}) Notification content. Must be set but can be an struct.
-// @param code(type=int) Notification code to use. Must be equal or greater than 0.
+// @param content(type=map[string]interface{}) Notification content. Must be set but can be any empty map.
+// @param code(type=int) Notification code to use. Must be greater than or equal to 0.
 // @param persistent(type=bool) Whether to record this in the database for later listing.
 // @return error(error) An optional error value if an error occurred.
-func (n *RuntimeGoNakamaModule) NotificationSendToAll(ctx context.Context, subject string, content map[string]interface{}, code int, persistent bool) error {
+func (n *RuntimeGoNakamaModule) NotificationSendAll(ctx context.Context, subject string, content map[string]interface{}, code int, persistent bool) error {
 	if subject == "" {
 		return errors.New("expects subject to be a non-empty string")
 	}
@@ -1685,7 +1685,7 @@ func (n *RuntimeGoNakamaModule) NotificationSendToAll(ctx context.Context, subje
 		CreateTime: createTime,
 	}
 
-	return NotificationSendToAll(ctx, n.logger, n.db, n.router, not)
+	return NotificationSendAll(ctx, n.logger, n.db, n.tracker, n.router, not)
 }
 
 // @group wallets