+++ server/runtime_lua_nakama.go
@@ -220,7 +220,7 @@ func (n *RuntimeLuaNakamaModule) Loader(l *lua.LState) int {
 		"match_signal":                       n.matchSignal,
 		"notification_send":                  n.notificationSend,
 		"notifications_send":                 n.notificationsSend,
-		"notifications_send_to_all":          n.notificationSendToAll,
+		"notification_send_all":              n.notificationSendAll,
 		"wallet_update":                      n.walletUpdate,
 		"wallets_update":                     n.walletsUpdate,
 		"wallet_ledger_update":               n.walletLedgerUpdate,
@@ -4648,7 +4648,7 @@ func (n *RuntimeLuaNakamaModule) matchList(l *lua.LState) int {
 // @summary Send one in-app notification to a user.
 // @param userId(type=string) The user ID of the user to be sent the notification.
 // @param subject(type=string) Notification subject.
-// @param content(type=table) Notification content. Must be set but can be an struct.
+// @param content(type=table) Notification content. Must be set but can be an empty table.
 // @param code(type=number) Notification code to use. Must be equal or greater than 0.
 // @param sender(type=OptString, optional=true) The sender of this notification. If left empty, it will be assumed that it is a system notification.
 // @param persistent(type=OptBool, optional=true, default=false) Whether to record this in the database for later listing.
@@ -4872,13 +4872,13 @@ func (n *RuntimeLuaNakamaModule) notificationsSend(l *lua.LState) int {
 }
 
 // @group notifications
-// @summary Send one in-app notification to all users.
+// @summary Send an in-app notification to all users.
 // @param subject(type=string) Notification subject.
-// @param content(type=table) Notification content. Must be set but can be an struct.
-// @param code(type=number) Notification code to use. Must be equal or greater than 0.
+// @param content(type=table) Notification content. Must be set but can be an empty table.
+// @param code(type=number) Notification code to use. Must be greater than or equal to 0.
 // @param persistent(type=OptBool, optional=true, default=false) Whether to record this in the database for later listing.
 // @return error(error) An optional error value if an error occurred.
-func (n *RuntimeLuaNakamaModule) notificationSendToAll(l *lua.LState) int {
+func (n *RuntimeLuaNakamaModule) notificationSendAll(l *lua.LState) int {
 	subject := l.CheckString(1)
 	if subject == "" {
 		l.ArgError(1, "expects subject to be a non-empty string")
@@ -4904,7 +4904,7 @@ func (n *RuntimeLuaNakamaModule) notificationSendToAll(l *lua.LState) int {
 	senderID := uuid.Nil.String()
 	createTime := &timestamppb.Timestamp{Seconds: time.Now().UTC().Unix()}
 
-	not := &api.Notification{
+	notification := &api.Notification{
 		Id:         uuid.Must(uuid.NewV4()).String(),
 		Subject:    subject,
 		Content:    content,
@@ -4914,8 +4914,8 @@ func (n *RuntimeLuaNakamaModule) notificationSendToAll(l *lua.LState) int {
 		CreateTime: createTime,
 	}
 
-	if err := NotificationSendToAll(l.Context(), n.logger, n.db, n.router, not); err != nil {
-		l.RaiseError(fmt.Sprintf("failed to send notifications: %s", err.Error()))
+	if err := NotificationSendAll(l.Context(), n.logger, n.db, n.tracker, n.router, notification); err != nil {
+		l.RaiseError(fmt.Sprintf("failed to send notification: %s", err.Error()))
 	}
 
 	return 0